Initial work for separate click ignore lists
The idea for this change is to address potential
issues with websites where the terms of use prevent
ad clicking by banning the end user for such behaviour.

Currently the only way to prevent such a ban is to either:
1. Disable ad clicks completely
2. Disable ADN for the domain

This change proposes to create a separate list for
domains in which the ads would be hidden but not clicked.
The functionality would be similar to that of the current
ADN whitelist where use end user lists domains where
they want ads to be shown.